backend/requirements.txt
langgraph
langchain
openai
python-dotenv
requests
fpdf
flask
flask-cors